func(client):création de deux sous-composants pour TeamRow : un de consultation et un d'édition/création
